Optimise grant ID validation when the ID is unchanged
When the submitted value is the same as the stored value, do not
bother validating it. If it was once valid, it should remain valid,
even if too old. By avoiding the lookup, we also avoid any chance of
network errors.

When the form is submitted, update the stored ID (looking up the
agreement_at) only if it's not the same as what's already in the DB. And
perform the deletion only if the field wasn't already blank.

Change-Id: I6cb691fc287efd39cc95812fa7b2da84fbb0742d