Refactor grant ID code to use exceptions for errors
The current error messages are really specific to the special page
workflow, and are completely meaningless e.g. in an API context.

One way to fix that could be to use generic error messages here, and
then do a key-based replacement in the special page hook handler.
However, that is still brittle, as message keys could change, and in
general it's not advisable to use them to vary the behaviour.

Instead, introduce new checked exception classes to use when the grant
ID is invalid, and when the Fluxx request fails. Let these exceptions
bubble up until we get to the special page handlers, and choose the
messages there.

This will also help with caching invalid IDs but not network failures.

Change-Id: I8614c178857d518a97490a3ae6eafe0d4eedce99