Use WAN cache for Fluxx data
As pointed out in T350947#9659908, the existing caching doesn't actually
do much, because it's per-server. So, use WAN cache instead, which is
shared between all appservers.

Also increase the base TTL to 1 hour. This increases the chance of the
value actually being reused. For the token, the TTL is still bound to
whatever the Fluxx API says is the maximum possible TTL. Also, because
we only cache successful responses, this should not incur in cache
invalidation problems, assuming that if a grant ID is valid it'll always
remain valid.

Finally, add the pcTTL option to let WanObjectCache perform in-object
caching, without pointlessly hitting memcached when the value was cached
in the same request.

Bug: T350947
Change-Id: I3c29c4aa69c792786c50a4fbd009af50e97f4ad0