Community-defined page collections: Add parser and API for metadata
To properly support community-defined page collections, we also need
to define and fetch the collection metadata for each such collection.
These collection metadata should be available through an API, that
will be used by the Recommendation API to fetch the collection metadata
and serve them along with the recommended articles for translation.

This patch adds the logic to support the definition of such collection
metadata, through the usage of a custom HTML marker that will include
the metadata as HTML attributes, like below:
```
<page-collection
  name='My Page Collection'
  description='This is a page collection'
  source-language='en'
  target-languages='es|fr'
  end-date='2024-10-20'
></page-collection>
```

More specifically, this patch:
- Uses a parser hook to parse such HTML markers, extract the collection
metadata from its HTML attributes and store them as extension data.
- Adds a tracking category to the pages that include such HTML marker,
so that we can easily find/query such pages.
- Add ApiQueryPageCollectionsMetadata Action API module to expose these
collection metadata for the requested pages (if they include such
markers).

Bug: T373132
Change-Id: I5f4b0041a4d3281702f2005ee1485b49610f8554