Implement pagination for WikiProjects
Use PagerNavigationBuilder to provide standard pagination links, like
the other paginated lists. This forces us to reimplement some of the
pagination logic ourselves (e.g., checking for a previous/next page);
support is somewhat basic (e.g., capped limit, offset needs to be in the
list), but that should be fine for this first iteration.

Bug: T374836
Change-Id: I9ab0eb6156a32b8f6b4727721516f80791d5050a