Add full pagination support to WikiProjectFullLookup
Add support for listing the results in reverse, and refactor the code a
bit -- this is a bit more verbose but hopefully easier to follow given
the various edge cases and the math involved (which is prone to
off-by-one errors).

Add a test for the pagination part, to verify that the math is correct.

Bug: T374836
Change-Id: Ie678eabc64ccdf2b45c03f606a9d94d7e8924a4c