Use null placeholder for WikiProjects with no sitelinks
By not skipping these items, we guarantee that the return value
contains as many WikiProjects as possible while respecting the requested
limit and offset. This makes pagination easier, as we don't have to
worry about a whole result page being filtered out, similar to T375407.

Since these WikiProjects are not useful to the user (they were likely
deleted/unlinked on Wikidata after we ran the WDQS query), we do not
display them in the UI. We only use their QIDs for pagination.

Change-Id: Ifb676dab100f26779a1c4dd1556067cb252402a5