WikiProjectIDLookup: avoid WikiMap for getting the local server
I don't know why were using WikiMap in the first place. This code only
needs the canonical server for the local wiki, which is available via
$wgCanonicalServer (which is also what WikiMap and WikiReference use).

So, use that directly to avoid dealing with the scenario where the
current wiki isn't known to WikiMap. That *can* actually happen if
$wgConf is not configured (which is typical in single-wiki
environments) and if the current wiki is not in the `sites` table
(and the installer does not add it by default). For example, that is
the case in Wikimedia CI.

Change-Id: I9be01fbb3503287ea8044609d3be91f8f6103ae6