PageCollectionMetadataApi: don't parse pages
The 'pagecollectionsmetadata' API endpoint provides basic
information about page collections (name, description, etc).
It does so by retrieving metadata from the parser output that
was stored by the <page-collection> custom tag handler.

The API code, however, was getting the parser output by
explicitly parsing the page, which is slow for large pages
with templates and Lua modules, which is the case for
those collections generated from wikidata items.

This patch only tries to get the output from the parser cache.

Bug: T387370
Change-Id: I113c7c345b88d9d9f4fcd07224cee2a6bdfbb13e