CollaborationListHandler: drop last usage of OOUI and clean up
The code is creating a plain <p> element, so OOUI is unnecessary. Note
that this usage would also be allowed in transclusion, as the <p> tag
can be rendered without accessing the global theme; nonetheless, using
the Html library is sufficient.

Avoid loading OOUI icons, no longer needed now that we use Codex
(I9c7c0d178be81b3e).

Introduce a local variable for the header for readability of the
subsequent array declaration.

Rename the hook parameter to match the interface signature, and drop
`@inheritDoc` as there is no doc comment to inherit.

Bug: T390084
Change-Id: I25367d4e1beaf6e1276a7e8e55df6f6ca4ab8a26